テストの共通データをまとめるように修正
